refactor(bench): split metric policy comparison
